more exceptions treated

+++ src/main/java/dev/fernando/dscommerce/services/ProductService.java
@@ -1,16 +1,18 @@
 package dev.fernando.dscommerce.services;
 
-import java.util.Optional;
-
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import dev.fernando.dscommerce.dto.ProductDTO;
 import dev.fernando.dscommerce.entities.Product;
 import dev.fernando.dscommerce.repositories.ProductRepository;
+import dev.fernando.dscommerce.services.exceptions.DatabaseException;
 import dev.fernando.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -42,14 +44,25 @@ public class ProductService {
     
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = this.productRepository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        return new ProductDTO(this.productRepository.save(entity));
+        try {
+            Product entity = this.productRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            return new ProductDTO(this.productRepository.save(entity));
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Produto de id = %d não encontrado".formatted(id));
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        this.productRepository.deleteById(id);
+        if (!this.productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Produto de id = %d não encontrado".formatted(id));
+        }
+        try {
+            this.productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Produto de id = %d possui dependências, portanto, não é possível excluí-lo.".formatted(id));
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
